+++ stroom-core-client/src/main/java/stroom/index/client/presenter/IndexVolumeGroupEditPresenter.java
@@ -27,7 +27,7 @@ import stroom.index.shared.IndexVolumeFields;
 import stroom.index.shared.IndexVolumeGroup;
 import stroom.index.shared.IndexVolumeGroupResource;
 import stroom.index.shared.IndexVolumeResource;
-import stroom.node.client.NodeCache;
+import stroom.node.client.NodeManager;
 import stroom.query.api.v2.ExpressionOperator;
 import stroom.query.api.v2.ExpressionUtil;
 import stroom.svg.client.SvgPresets;
@@ -59,7 +59,7 @@ public class IndexVolumeGroupEditPresenter
     private final IndexVolumeStatusListPresenter volumeStatusListPresenter;
     private final Provider<IndexVolumeEditPresenter> editProvider;
     private final RestFactory restFactory;
-    private final NodeCache nodeCache;
+    private final NodeManager nodeManager;
 
     private final ButtonView newButton;
     private final ButtonView openButton;
@@ -76,12 +76,12 @@ public class IndexVolumeGroupEditPresenter
                                          final IndexVolumeStatusListPresenter volumeStatusListPresenter,
                                          final Provider<IndexVolumeEditPresenter> editProvider,
                                          final RestFactory restFactory,
-                                         final NodeCache nodeCache) {
+                                         final NodeManager nodeManager) {
         super(eventBus, view);
         this.volumeStatusListPresenter = volumeStatusListPresenter;
         this.editProvider = editProvider;
         this.restFactory = restFactory;
-        this.nodeCache = nodeCache;
+        this.nodeManager = nodeManager;
 
         newButton = volumeStatusListPresenter.getView().addButton(SvgPresets.NEW_ITEM);
         openButton = volumeStatusListPresenter.getView().addButton(SvgPresets.EDIT);
@@ -105,7 +105,7 @@ public class IndexVolumeGroupEditPresenter
         registerHandler(deleteButton.addClickHandler(event -> delete()));
         registerHandler(rescanButton.addClickHandler(event -> {
             final Rest<Boolean> rest = restFactory.create();
-            nodeCache.listAllNodes(nodeNames ->
+            nodeManager.listAllNodes(nodeNames ->
                             nodeNames.forEach(nodeName ->
                                     rest
                                             .onSuccess(response -> volumeStatusListPresenter.refresh())